Factor out semtree.py.
